Initial working release with cabal file.
It's neither nice nor beautiful, but it gets the job done. This is primarily
intended to get a tarball out of a Star Control II 3DO CD image with fixed
timestamps and just the video files we're interested in.

The reason I'm using fixed timestamps is that we need to get the same SHA256,
regardless of the current time and other weird natural disasters.

Signed-off-by: aszlig <aszlig@redmoonstudios.org>